requirement.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.1.0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
batchgenerators @ git+https://github.com/MIC-DKFZ/batchgenerators.git@ea0189973fbb6f0b2dbf7018d3c3b1cdab5dfc5f
beautifulsoup4==4.10.0
black==21.12b0
bleach==4.1.0
blessings==1.7
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
clip @ git+https://github.com/openai/CLIP.git@40f5484c1c74edd83cb9cf687c6ab92b28d8b656
ConfigArgParse==1.5.3
configparser==5.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
deepspeed==0.6.7
defusedxml==0.7.1
dicom2nifti==2.3.2
docker-pycreds==0.4.0
dominate==2.6.0
easydict==1.9
elasticdeform==0.4.9
entrypoints==0.4
filelock==3.6.0
fonttools==4.28.5
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gitdb==4.0.9
GitPython==3.1.25
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gpustat==0.6.0
grpcio==1.44.0
h5py==3.6.0
hjson==3.0.2
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-server==1.15.6
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.2
linecache2==1.0.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
MedPy==0.4.0
mistune==0.8.4
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nibabel==3.2.2
ninja==1.10.2.3
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.0.0
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parmap==1.5.3
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.1
psutil==5.9.0
ptyprocess==0.7.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pydantic==1.9.1
pyDeprecate==0.3.1
pydicom==2.2.2
pyparsing==3.0.6
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.0.8
pytz==2021.3
PyWavelets==1.2.0
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
qudida==0.0.4
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
SecretStorage==2.3.1
Send2Trash==1.8.0
sentry-sdk==1.5.1
shortuuid==1.0.8
SimpleITK==2.1.1
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
subprocess32==3.5.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.3
torchmetrics==0.7.2
tqdm==4.62.3
traceback2==1.4.0
traitlets==5.1.1
typed-ast==1.5.1
typing_extensions==4.0.1
unittest2==1.1.0
urllib3==1.26.8
visdom==0.1.8.9
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.7.0
